+++ docs/changelog.md
@@ -2,6 +2,32 @@
 
 > 记录项目每个版本的主要更新内容。
 
+## v1.2.22 · 2025-10-23
+
+> 取货码: 定长8位数字字符串，生成后允许客户在售卖机APP上进行取货，取货数量 1-3个，取货码的订单是免费的。
+
+取货码生成可通过接口生成，管理面板生成功能会在约定时间内关闭。
+
+[GraphQL 获取取货码](/GraphQL/pickup_code) - 适用于 APP
+
+### GRPC 文档
+
+- [20. 获取当前设备取货码(未过期)](/grpc/pickup_codes)
+- [21. 通过接口生成取货码](/grpc/create_pickup_codes)
+
+### MQTT 订阅 Subscribe
+
+- [13. 取货码生成后推送](/mqtt/subscribe/client.pickup_codes.create)
+- [14. 取货码删除后推送](/mqtt/subscribe/client.pickup_codes.delete)
+
+### MQTT 发布 Publish
+
+- [10. 取货码消费后推至服务端](/mqtt/publish/client.pickup_codes.put)
+
+### 接口推送文档
+
+- [批量上传订单](/standalone/batch_upload_orders) - 需要修改 `pickup_code_id` 取货码 ID
+
 ## v1.2.21 · 2025-10-22
 
 - 修改  `_at` 时间类字段，根据 `machine.timezone` 时区转换偏移量
@@ -12,7 +38,7 @@
 | `updated_at` | 更新时间 | datetime(3) | UTC |
 | `expired_at` | 过期时间 | datetime(3) | UTC |
 
-即 `graphQL`, `GRPC` 所有接口请求返回的 `_at` 字段值都是转换时区后的时间，`秒级,毫秒级时间戳` 亦然。
+即 `GraphQL`, `GRPC` 所有接口请求返回的 `_at` 字段值都是转换时区后的时间，`秒级,毫秒级时间戳` 亦然。
 
 ## v1.2.20 · 2025-10-15